fix: CodeRabbitレビュー指摘への対応
- spawner.ts: SpawnOptions.env を Record<string, string | undefined> に変更して process.env のスプレッド代入に対応
- windows.ts: load/delete から validateProfile を削除（save のみに限定）、PowerShell スポーン時に process.env を継承して PATH/SystemRoot を保持
- linux.ts: load/delete から validateProfile を削除（save のみに限定）
- macos.ts: load/delete から validateProfile を削除（save のみに限定）
- テストヘルパーを _keychain-test-helpers.ts に集約して重複排除
- keychain-windows.test.ts: COS_TARGET アサーションを厳格化（$env:COS_TARGET、env["COS_TARGET"] の値を直接検証）
