+++ src/infra/keychain/windows.ts
@@ -66,7 +66,6 @@ export class WindowsKeychainStore implements TokenStore {
   }
 
   async load(profile: string): Promise<string | null> {
-    validateProfile(profile)
     // cmdkey は直接パスワードを読み出せないため PowerShell + CredentialManager モジュールを試みる。
     // profile 値は文字列補間を避けるため環境変数 COS_TARGET 経由で渡す。
     const script = `
@@ -82,7 +81,8 @@ export class WindowsKeychainStore implements TokenStore {
       proc = this.spawn(["powershell", "-NoProfile", "-Command", script], {
         stdout: "pipe",
         stderr: "pipe",
-        env: { [ENV_TARGET]: `${SERVICE}:${profile}` },
+        // process.env を展開して PATH / SystemRoot など Windows 必須変数を引き継ぐ
+        env: { ...process.env, [ENV_TARGET]: `${SERVICE}:${profile}` },
       })
     } catch (e) {
       if (isENOENT(e)) throw new Error(POWERSHELL_NOT_FOUND_MESSAGE)
@@ -101,7 +101,6 @@ export class WindowsKeychainStore implements TokenStore {
   }
 
   async delete(profile: string): Promise<void> {
-    validateProfile(profile)
     let proc: SubprocessLike
     try {
       proc = this.spawn(["cmdkey", `/delete:${SERVICE}:${profile}`], {
